requirements.txt
# Web Framework & API
# ===================
Flask >=2.3.2
flask-cors >=3.0.10
waitress >=3.0.2
pydantic>=2.0.0

# Database & Data Storage
# =======================
trino
boto3
fsspec
s3fs 
fastparquet
pyarrow

# Core ML & Data Processing
# =========================
numpy>=1.26.4
pandas>=2.2.1
scipy>=1.13.0
scikit-learn>=1.4.2
joblib >=1.3.2

# Machine Learning Models
# =======================
lightgbm>=3.3.6
xgboost>=2.0.3
imbalanced-learn>=0.12.2
numpyro>=0.10.0
shap>=0.44.0

# Model Serving & ONNX
# =====================
skl2onnx>=1.18.0
onnx>=1.18.0
onnxruntime>=1.19.0
onnxsim>=0.4.36
onnxmltools>=1.14.0
onnxconverter_common>=1.14.0

# SQL & NLP Parsing
# =================
sqlglot>=10.0.0
sqlparse>=0.5.0
sqllineage>=1.5.0
sql-metadata>=1.1.0
spacy>=3.7.2

# System & Utilities
# ==================
psutil
pyyaml>=6.0.1
python-json-logger>=2.0.7

# Testing & Development
# =====================
pytest
pytest-mock
pytest-cov
